+++ src/app/locales/en/components/content-privacy-policy.js
@@ -1,4 +1,3 @@
 export default {
   privacyPolicy: 'Privacy policy',
-  backToMainPage: 'Back to main page',
 };